ver 0.1.0
Simpie has now been compiled to an alpha state executable. Simpie can now be downloaded and executed on almost every windows computer. In following updated MacOS support will follow, and after that Linux support.A funny little icon was also added to the repository. This will be the icon of Simpie for now. The original png was also included in this update.