ver 0.1.4
Major update, you can now read the devblog from within simpie. So if you are interested give it a try! With this update there are a lot of new features so go over to the devblog to read more.
In short:
version command in settings
simpie guesses what you typed if you had a typo
simpie will check on startup if it is up to date
devblog command
lookup command / google / search
and much more.