ver 0.1.6
SimpieBot has now been compiled for linux. Starting the bot on linux is not as easy as just double clicking it open the readme file. Most of the commands will work in the linux dist of simpiebot. More info is available in the devblog. Thanks for using SimpieBot and have fun.